Major E0308 type mismatch resolution using 25 parallel subagents
- Reduced compilation errors from 234 to 202 (-32 errors, 14% improvement)
- Fixed 30 of 78 E0308 type mismatch errors (38% improvement) through systematic type conversions
- Resolved interface registry HashMap type mismatches and method signature compatibility
- Fixed LLVM API BasicValueEnum conversions and Result type handling patterns
- Completed Arc/RwLock thread safety conversions in stdlib and runtime modules
- Fixed numerous String vs &str type mismatches across codegen modules
- Corrected struct field access patterns and private field access issues

Key modules improved:
- Interface type assertion modules (path visualization, debug, result integration)
- LLVM codegen modules (concurrency, property access, enhanced dispatch)
- Core interface registry and type system modules
- Runtime and stdlib modules for thread safety compliance
- Parser and error handling modules for type consistency

+++ src/error/type_assertion_error.rs
@@ -184,7 +184,7 @@ impl From<crate::error::Error> for TypeAssertionError {
                 }
                 
                 let location_opt = assertion_error.location().map(|loc| {
-                    SourceLocation::new(loc.line as usize, loc.column as usize)
+                    SourceLocation::new(loc.line, loc.column)
                 });
                 
                 if let Some(loc) = location_opt {
